Preparation to refactor Just a checkpoint for refactoring Expr.hpp

+++ src/Headers/bisayaPP.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <string>
 
 #include "Token.hpp"
@@ -9,8 +10,8 @@ namespace BisayaPP {
 void run(const std::string& source);
 void runFile(const char* path);
 void runPrompt();
-void error(int line, const std::string& message);
-void report(int line, const std::string& where, const std::string& message);
+void error(size_t line, const std::string& message);
+void report(size_t line, const std::string& where, const std::string& message);
 void error(const Token& token, const std::string& message);
 
 }  // namespace BisayaPP
